requirements.txt
pandas
scikit-learn
plotly
python-dotenv
numpy
flask
requests
flask-cors
